Answer security questions
Extract from the new option definition:

Some web sites use "confirmation questions" that, supposedly, only
you know the answer to.  Unfortunately, these are often times
questions that can be researched by others or they can be questions
that you have forgotten the answers to or may have multiple answers
for.  The net result is that you are locked out.  This option makes
it easy to get consistent answers to these questions and have these
answers be different for every web site, just like your password.

Providing this option will cause the argument to be merged into the
hash source (changing the resulting password).  Exactly 12 letters
will be extracted from the hash and converted to lower case.  The
string argument to this option should be the last word or two from
the question, yielding an easy-to-remember way of obtaining a
consistent answer to these inscrutable questions.

You will need to update your confirmation question answers when
you update your password seed.