keep track of domains
When the number of passwords approaches many tens of logins,
it becomes unwieldy to track which domains have passwords.
With the new --domain option, this program will keep track
of a domain name and a last accessed date.  The saved
domain name should not reveal the domain transform.  If
you do, then you have effectively stored the password id
in plain text.  Not a good idea.

* bootstrap.conf: do not pull static variables into forward
  declarations.

* src/cfg-file (find_cfg_name): new function
  (find_dom_file): new function

* src/gnu-pw-mgr.c (load_file): new funtion to load all text
  (list_domains): print the contents of the domain file
  (write_dom_file): (re)write the contents of the domain text
  (insert_domain): either update the day-of-access or add a
  new entry to the domain text
  (proc_dom_opts): new function to handle the options

* src/gpw-str.def: add entries for domain text
* src/opts.def (domain): new option