add a 3 characters in sequence password fix
Chase Bank has added the password restriction that no three characters
may be in sequence. So look for that and fix it, if need be.

Let people know when they last modified a password.
Mark the day when the pbkdf2 count is modified.

The confirmation value printed out is a hash based on the confirmation
string and the current password. So, if the password changes, then
so does the confirmation string. Not ideal. This change restricts
the hashing source to just the password id and the confirmation
string.

update copyright year to 2018

* bootstrap.conf (init_srcdir): no handler for --rehash, so don't count it
* doc/intro.texi: clarification
* src/gnu-pw-mgr.c (set_confirm_value): off by 1
  (print_pwid): call new function when header not suppressed
  (print_pw_header): function to cleanup logic
  (set_confirm_value): new function for doing this work.
  (adjust_pw): no longer called to do confirmation fixups
  (get_dft_pw): call set_confirm_value if --confirm is set
  (get_pbkdf2_pw): likewise
* src/opts.def (rehash): new option that triggers --use-pbkdf2
  (pbkdf2): mark as not for use
  (detail): doc cleanup
  (SECONDS_IN_DAY): new define
  (pbkdf2_date): new global
  (confirm): update the description
* src/sort-opts.def: doc clarification
* src/gpw-str.def: several new strings
* src/pw-opts.c (day_to_string): convert days since 1970 to YYYY-MM-DD
  (remove_opt): adjust for the fact that the pwtag for pbkdf2 options
  now contain a ``date="..."'' field
  (update_pwid_opts): insert that field
* tests/base.test: add to and adjust tests
* doc/warnings.texi: update